recipe-app/requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.5
django-environ==0.12.0
fonttools==4.55.3
gunicorn==23.0.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
